first pass refactoring done
