+++ NavigationPolicy/Policies/OpenInNewTabPolicy.swift
